msmnile_gvmq: Set persist.bluetooth.enablenewavrcp as false
The property "persist.bluetooth.enablenewavrcp" is introduced in AOSP.

See commit e63f6d6bda16bd94d43537fc5db754a103c6a757

(1) If the property is set as true, it indicates that AVRCP(TG) is enabled.

(2) If the property is set as false, it indicates that AVRCP(CT) is enabled.

In Fluoride Bluetooth stack, the default value for the property is true. This is
valid with Mobile SP.

However in Automotive SP, AVRCP(CT) is enabled in Car UI.

So the property should be set as false.

Change-Id: I0d5c4e7fd12463799ab4359d7f724b5d1a8e1144