autogvm: Enable compilation of init_boot partition
Devices launching with Android 13 or higher and shipping with
Android 13 must have an init_boot static partition. The generic
ramdisk is no longer be contained in the boot.img file that goes
in the boot partition. This generic ramdisk will be placed in a
new init_boot.img to be installed in a new init_boot partition.

Change-Id: I08c043b056c4023099c8bca0a1d5fedc3b2f9927
Signed-off-by: Singa Reddy Dasari <quic_singredd@quicinc.com>